+++ routes/api/listing-routes.js
@@ -2,7 +2,15 @@ const express = require('express');
 const router = express.Router({ mergeParams: true });
 const { check, oneOf, query } = require('express-validator');
 const { advancedResults, checkInputError, protect, authorise, mapFilenameToLocation } = require('../../middleware');
-const { DATETIME_REGEX, NO_FIELD_UPDATED_MSG, INVALID_FIELD_MSG, INVALID_EMAIL_MSG, INVALID_BOOLEAN_MSG, INVALID_TIMESTAMP_MSG, uploadFile } = require('../../utils');
+const {
+    DATETIME_REGEX,
+    NO_FIELD_UPDATED_MSG,
+    INVALID_FIELD_MSG,
+    INVALID_EMAIL_MSG,
+    INVALID_BOOLEAN_MSG,
+    INVALID_TIMESTAMP_MSG,
+    uploadFile,
+} = require('../../utils');
 
 // import controllers here
 const {
@@ -52,7 +60,7 @@ const validateCreateListingFields = [
     check('title', INVALID_FIELD_MSG('title')).trim().notEmpty(),
     check('category', INVALID_FIELD_MSG('category')).trim().notEmpty(),
     check('listing_url', INVALID_FIELD_MSG('listing URL')).optional().isURL(),
-    check('listing_email', INVALID_EMAIL_MSG('listing email')).notEmpty().isEmail(),
+    check('listing_email', INVALID_EMAIL_MSG).notEmpty().isEmail(),
     check('is_published', INVALID_BOOLEAN_MSG('is_published')).optional().isBoolean(),
     check('start_date', INVALID_TIMESTAMP_MSG('start date')).optional().matches(DATETIME_REGEX),
     check('end_date', INVALID_TIMESTAMP_MSG('end date')).optional().matches(DATETIME_REGEX),
@@ -84,7 +92,7 @@ const validateUpdateListingFields = [
     check('title', INVALID_FIELD_MSG('title')).optional().trim().notEmpty(),
     check('category', INVALID_FIELD_MSG('category')).optional().trim().notEmpty(),
     check('listing_url', INVALID_FIELD_MSG('listing URL')).optional().isURL(),
-    check('listing_email', INVALID_EMAIL_MSG('listing email')).optional().isEmail(),
+    check('listing_email', INVALID_EMAIL_MSG).optional().isEmail(),
     check('is_published', INVALID_BOOLEAN_MSG('is_published')).optional().isBoolean(),
     check('start_date', INVALID_TIMESTAMP_MSG('start date')).optional().matches(DATETIME_REGEX),
     check('end_date', INVALID_TIMESTAMP_MSG('end date')).optional().matches(DATETIME_REGEX),