+++ utils/inputExceptionMsg.js
@@ -1,19 +1,16 @@
-exports.INVALID_FIELD_MSG = field => `Please include a valid ${field}`;
+exports.INVALID_FIELD_MSG = (field) => `Please include a valid ${field}`;
 
-exports.INVALID_EXISTING_MSG = field => `Please enter your ${field}`;
+exports.INVALID_EXISTING_MSG = (field) => `Please enter your ${field}`;
 
-exports.INVALID_ALPHA_SPACE_MSG = field =>
-  `Please include a valid ${field} with alphabets and spaces only`;
+exports.INVALID_ALPHA_SPACE_MSG = (field) => `Please include a valid ${field} with alphabets and spaces only`;
 
-exports.INVALID_BOOLEAN_MSG = field =>
-  `Please include a boolean value for ${field}`;
+exports.INVALID_BOOLEAN_MSG = (field) => `Please include a boolean value for ${field}`;
 
-exports.INVALID_TIMESTAMP_MSG = field =>
-  `Please include a valid timestamp (of string type) for ${field}`;
+exports.INVALID_TIMESTAMP_MSG = (field) => `Please include a valid timestamp (of string type) for ${field}`;
 
 exports.INVALID_EMAIL_MSG = 'Please include a valid email';
 
 exports.INVALID_PASSWORD_MSG =
-  'Please enter a password with 6 or more characters';
+    'Please enter a password with 8-25 characters, at least 1 uppercase letter, 1 lowercase letter and 1 special character';
 
 exports.NO_FIELD_UPDATED_MSG = 'At least one field must be updated';